Lint: split checks into multiple modules
